local database implemented for future orders

+++ app/src/main/java/com/example/smartlab/utils/extensions.kt
@@ -8,4 +8,4 @@ import androidx.datastore.preferences.preferencesDataStore
 val Context.dataStore: DataStore<Preferences> by preferencesDataStore(name = "smartlab")
 
 
-fun Context.dpToPx(dp: Int): Int = (dp * resources.displayMetrics.density).toInt()
+fun Context.dpToPx(dp: Double): Int = (dp * resources.displayMetrics.density).toInt()
